requirements.txt
streamlit
google-genai>=0.3.0
pandas